refactor: centralize popularity calc and text utils
